Added a repository layer for db
Yes, I know that DbContext and DbSet act as a unit of work and repository layer respectively, but adding a repository layer allows me to use PostgreSQL specific functions. A lot more things could be done to optimize the layer (like adding a unit of work) but this would suffice for now.